pyproject.toml
[project]
name = "ronnykraitman"
version = "0.1.0"
description = "This is Ronny Kraitman's personal repo"
readme = "README.md"
requires-python= "==3.12.10"
dependencies = [
    "anthropic>=0.49.0",
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,ollama,openai]>=0.4.9.2",
    "langchain-openai>=0.3.9",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "playwright>=1.51.0",
    "plotly>=6.0.1",
    "polygon-api-client>=1.14.5",
    "psutil>=7.0.0",
    "pypdf>=5.4.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "semantic-kernel>=1.25.0",
    "streamlit"
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.poetry]
package-mode = false